@@ -26,16 +26,19 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.4 || ^8.0",
-        "behat/behat": "^3.0",
-        "symfony/messenger": "^4.4 || ^5.0",
-        "symfony/serializer": "^4.4 || ^5.0"
+        "php": ">=8.1",
+        "behat/behat": "^3.6.0",
+        "symfony/messenger": "^6.3",
+        "symfony/serializer": "^6.3",
+        "illuminate/collections": "^10.16"
     },
     "require-dev": {
-        "phpstan/phpstan": "^0.12",
-        "phpunit/phpunit": "^9.3",
-        "slevomat/coding-standard": "^7.0",
-        "squizlabs/php_codesniffer": "^3.6"
+        "phpstan/phpstan": "^1.10.12",
+        "phpunit/phpunit": "^10.2.6",
+        "slevomat/coding-standard": "^8.7.1",
+        "squizlabs/php_codesniffer": "^3.6",
+        "rector/rector": "^0.17.7",
+        "zenstruck/messenger-test": "^1.7"
     },
     "autoload": {
         "psr-4": {
@@ -52,7 +55,7 @@
         "code-style": "./vendor/bin/phpcs",
         "code-style-fix": "./vendor/bin/phpcbf",
         "phpunit": "./vendor/bin/phpunit",
-        "dev-checks": [
+        "test": [
             "composer validate",
             "@phpstan",
             "@code-style",
