@@ -38,7 +38,7 @@ class ArraySimilarTraitTest extends TestCase
     {
         $result = $this->isArraysSimilar(
             [
-                'time' => '~'.self::ATOM_DATETIME_PATTERN,
+                'time' => '~' . self::ATOM_DATETIME_PATTERN,
                 'foo' => 1,
             ],
             [
@@ -53,6 +53,11 @@ class ArraySimilarTraitTest extends TestCase
 
     /**
      * @dataProvider variableFieldsFailProvider
+     *
+     * @var array<mixed>$expected
+     * @var array<mixed>$actual
+     * @var array<string>$variableFields
+     * @var array<string,string>$actual
      */
     public function testVariableFieldsFail(
         array $expected,
@@ -65,7 +70,7 @@ class ArraySimilarTraitTest extends TestCase
         self::assertFalse($result);
     }
 
-    public function variableFieldsFailProvider(): iterable
+    public static function variableFieldsFailProvider(): iterable
     {
         yield '#1: Value type different from string' => [
             ['a' => 1],
@@ -105,7 +110,7 @@ class ArraySimilarTraitTest extends TestCase
             ['date' => '{datetime_atom}'],
             ['date' => (new \DateTimeImmutable())->format(\DateTimeInterface::ATOM)],
             ['date'],
-            ['datetime_atom' => '/'.self::ATOM_DATETIME_PATTERN.'/'],
+            ['datetime_atom' => '/' . self::ATOM_DATETIME_PATTERN . '/'],
         );
 
         self::assertTrue($result);
